+++ src/tkc/object_date_time.c
@@ -22,6 +22,7 @@
 #include "tkc/mem.h"
 #include "tkc/value.h"
 #include "tkc/utils.h"
+#include "tkc/named_value.h"
 #include "object_date_time.h"
 
 static ret_t object_date_time_on_destroy(tk_object_t* obj) {
@@ -62,6 +63,11 @@ static ret_t object_date_time_set_prop(tk_object_t* obj, const char* name, const
     }
     case 's': {
       date_time_set_second(o->dt, value_uint32(v));
+      break;
+    }
+    case 'w': {
+      /*日期确定后，wday也确定了，支持设置wday方便测试检查结果*/
+      o->dt->wday = value_uint32(v);
       break;
     }
     default: {
@@ -114,13 +120,60 @@ static ret_t object_date_time_get_prop(tk_object_t* obj, const char* name, value
   return ret;
 }
 
-static const object_vtable_t s_object_date_time_vtable = {.type = "object_date_time",
-                                                          .desc = "object_date_time",
-                                                          .size = sizeof(object_date_time_t),
-                                                          .is_collection = FALSE,
-                                                          .on_destroy = object_date_time_on_destroy,
-                                                          .get_prop = object_date_time_get_prop,
-                                                          .set_prop = object_date_time_set_prop};
+static ret_t object_date_time_foreach_prop(tk_object_t* obj, tk_visit_t on_prop, void* ctx) {
+  named_value_t nv;
+  ret_t ret = RET_OK;
+  object_date_time_t* o = OBJECT_DATE_TIME(obj);
+  return_value_if_fail(on_prop != NULL, RET_BAD_PARAMS);
+  return_value_if_fail(o != NULL && o->dt != NULL, RET_BAD_PARAMS);
+
+  nv.name = OBJECT_DATE_TIME_PROP_YEAR;
+  value_set_uint32(&(nv.value), o->dt->year);
+  ret = on_prop(ctx, &nv);
+  return_value_if_fail(ret == RET_OK, RET_FAIL);
+
+  nv.name = OBJECT_DATE_TIME_PROP_MONTH;
+  value_set_uint32(&(nv.value), o->dt->month);
+  ret = on_prop(ctx, &nv);
+  return_value_if_fail(ret == RET_OK, RET_FAIL);
+
+  nv.name = OBJECT_DATE_TIME_PROP_MINUTE;
+  value_set_uint32(&(nv.value), o->dt->minute);
+  ret = on_prop(ctx, &nv);
+  return_value_if_fail(ret == RET_OK, RET_FAIL);
+
+  nv.name = OBJECT_DATE_TIME_PROP_DAY;
+  value_set_uint32(&(nv.value), o->dt->day);
+  ret = on_prop(ctx, &nv);
+  return_value_if_fail(ret == RET_OK, RET_FAIL);
+
+  nv.name = OBJECT_DATE_TIME_PROP_HOUR;
+  value_set_uint32(&(nv.value), o->dt->hour);
+  ret = on_prop(ctx, &nv);
+  return_value_if_fail(ret == RET_OK, RET_FAIL);
+
+  nv.name = OBJECT_DATE_TIME_PROP_SECOND;
+  value_set_uint32(&(nv.value), o->dt->second);
+  ret = on_prop(ctx, &nv);
+  return_value_if_fail(ret == RET_OK, RET_FAIL);
+
+  nv.name = OBJECT_DATE_TIME_PROP_WDAY;
+  value_set_uint32(&(nv.value), o->dt->wday);
+  ret = on_prop(ctx, &nv);
+  return_value_if_fail(ret == RET_OK, RET_FAIL);
+
+  return RET_OK;
+}
+
+static const object_vtable_t s_object_date_time_vtable = {
+    .type = "object_date_time",
+    .desc = "object_date_time",
+    .size = sizeof(object_date_time_t),
+    .is_collection = FALSE,
+    .on_destroy = object_date_time_on_destroy,
+    .foreach_prop = object_date_time_foreach_prop,
+    .get_prop = object_date_time_get_prop,
+    .set_prop = object_date_time_set_prop};
 
 tk_object_t* object_date_time_create(void) {
   tk_object_t* o = NULL;