+++ docs/changes.md
@@ -3,6 +3,7 @@
 2022/03/19
   * 完善fscript函数bit\_not
   * 完善fargs\_get\_data\_and\_size
+  * 完善object\_date\_time支持foreach。
 
 2022/03/18
   * 完善 tk_\cond\_var\_wait（感谢兆坤提供补丁）。